LPS-54010 Convert ItemSelectorViewRenderer into an interface to make API cleaner and easier to maintain.
